feat: wrapper around bash's builtin exit
The wrapper allows passing additional arguments that are interpreted
as arguments to a printf call and can be used to print a message
before exiting.

The wrapper also provides a way to perform certain actions when a
well known error code is used. For now this includes calling a
usage() function (if it exists) when the error code 64 (EX_USAGE)
is used.

Signed-off-by: shypa <shypa@devyard.org>