exit: fix exit on error when shifting arguments
Bash's builtin `shift` will return a non-zero exit status if called from
a context where "$@" is empty. This means that for environments running
with `set -e` or `set -o errexit` and which call `exit` without
arguments, our implementation used to exit with an error
unconditionnally.

This commits fixes that by hiding/ignoring the return status of `shift`
in `exit`.

Signed-off-by: shypa <shypa@devyard.org>