colors: prefix/namespace function with 'colors.'
Note that `_ps1_classic()` was updated to reuse the color list defined
in `sources.d/colors:colors`. As such, the color ordering now depends
on the ordering of the keys in the `colors` associative array.

Also note that only the stability of that ordering is required, which
appears to be the case.